+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -1084,6 +1084,7 @@ class HubTest {
         sentryTracer.finish()
         sut.captureTransaction(SentryTransaction(sentryTracer), null)
         verify(mockClient, never()).captureTransaction(any(), any(), any())
+        verify(mockClient, never()).captureTransaction(any(), any(), any(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -1099,7 +1100,43 @@ class HubTest {
         val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
         sentryTracer.finish()
         val traceState = sentryTracer.traceState()
-        verify(mockClient).captureTransaction(any(), eq(traceState), any(), eq(null), eq(null))
+        verify(mockClient).captureTransaction(any(), eq(traceState), any(), eq(null), anyOrNull())
+    }
+
+    @Test
+    fun `when startTransaction and profiling is enabled, transaction is profiled only if sampled`() {
+        val mockTransactionProfiler = mock<ITransactionProfiler>()
+        val hub = generateHub {
+            it.isProfilingEnabled = true
+            it.setTransactionProfiler(mockTransactionProfiler)
+        }
+        // Transaction is not sampled, so it should not be profiled
+        val contexts = TransactionContext("name", "op", false)
+        val transaction = hub.startTransaction(contexts)
+        transaction.finish()
+        verify(mockTransactionProfiler, never()).onTransactionStart(anyOrNull())
+        verify(mockTransactionProfiler, never()).onTransactionFinish(anyOrNull())
+
+        // Transaction is sampled, so it should be profiled
+        val sampledContexts = TransactionContext("name", "op", true)
+        val sampledTransaction = hub.startTransaction(sampledContexts)
+        sampledTransaction.finish()
+        verify(mockTransactionProfiler).onTransactionStart(anyOrNull())
+        verify(mockTransactionProfiler).onTransactionFinish(anyOrNull())
+    }
+
+    @Test
+    fun `when startTransaction and is sampled but profiling is disabled, transaction is not profiled`() {
+        val mockTransactionProfiler = mock<ITransactionProfiler>()
+        val hub = generateHub {
+            it.isProfilingEnabled = false
+            it.setTransactionProfiler(mockTransactionProfiler)
+        }
+        val contexts = TransactionContext("name", "op", true)
+        val transaction = hub.startTransaction(contexts)
+        transaction.finish()
+        verify(mockTransactionProfiler, never()).onTransactionStart(anyOrNull())
+        verify(mockTransactionProfiler, never()).onTransactionFinish(anyOrNull())
     }
 
     @Test
@@ -1111,7 +1148,7 @@ class HubTest {
         val sut = Hub(options)
         val mockClient = mock<ISentryClient>()
         sut.bindClient(mockClient)
-        whenever(mockClient.captureTransaction(anyOrNull(), anyOrNull(), anyOrNull(), anyOrNull())).thenReturn(SentryId())
+        whenever(mockClient.captureTransaction(anyOrNull(), anyOrNull(), anyOrNull(), anyOrNull(), anyOrNull())).thenReturn(SentryId())
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
         sentryTracer.finish()
@@ -1130,7 +1167,7 @@ class HubTest {
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
         sut.captureTransaction(SentryTransaction(sentryTracer), null)
-        verify(mockClient, never()).captureTransaction(any(), any(), any(), eq(null))
+        verify(mockClient, never()).captureTransaction(any(), any(), any(), eq(null), anyOrNull())
     }
 
     @Test
@@ -1146,7 +1183,7 @@ class HubTest {
         val sentryTracer = SentryTracer(TransactionContext("name", "op", false), sut)
         sentryTracer.finish()
         val traceState = sentryTracer.traceState()
-        verify(mockClient, never()).captureTransaction(any(), eq(traceState), any(), eq(null))
+        verify(mockClient, never()).captureTransaction(any(), eq(traceState), any(), eq(null), anyOrNull())
     }
     //endregion