+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Feat: Add Android profiling traces #1897
+* Feat: Add Android profiling traces #1897 and its tests #1949
 
 ## 5.6.2